update n_sel_files_ using g_list_length()

+++ src/ptk/ptk-file-browser.hxx
@@ -103,7 +103,7 @@ struct PtkFileBrowser
     vfs::dir dir_{nullptr};
     GtkTreeModel* file_list_{nullptr};
     i32 max_thumbnail_{0};
-    i32 n_sel_files_{0};
+    u64 n_sel_files_{0};
     u64 sel_size_{0};
     u64 sel_disk_size_{0};
     u32 sel_change_idle_{0};
